+++ src/world/mod.rs
@@ -18,7 +18,7 @@ pub struct World {
     pub settings: Settings,
 
     // variables
-    pub(crate) time: f64, // current time in the simulation
+    pub(crate) time: f32, // current time in the simulation
     pub(crate) population_size: usize, // current population size of entities
     pub(crate) ligands_count: usize, // current count of ligands in the world
